set dummy minimal required version in dependencies
The plugin depends on a number of Perl modules. For some I don't know what exact
version is the minimal requirement (and I have no clue on how I cound determine
that easily). And the "configure" thingy apparently doesn't like ">=0", so we
set it to "=>0.1" which it hopefully likes more.